refact: extract shared topicwritetest helpers for direct write tests
